Color support for dvipng-based previews.
Trying to do this at the tex/DVI level was a bust. This approach uses
ghostscript (as run by dvipng) to set the image colors when creating them. This
is fine for basic color matching, but to match each snippet to its surrounding
colors (such as in headlines) will require multiple dvipng processes and be
challenging.